pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.8",
    "openai>=1.84.0",
    "pymupdf>=1.26.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "python-markdown-math>=0.9",
    "weasyprint>=65.1",
]
